Cleanup, some comments and test reorg
